Move layout to components dir
